[DF-11072] fix remark
Former-commit-id: f34ebc0f8f09be0f7d66a51067d72b238ed66cb7

+++ Sources/Internal/Platform/SystemTimer.cpp
@@ -272,7 +272,7 @@ uint64 SystemTimer::GetSystemTime()
 {
     auto sysTime = std::chrono::system_clock::now();
     time_t sysTimeT = std::chrono::system_clock::to_time_t(sysTime);
-    return static_cast<uint32>(sysTimeT);
+    return static_cast<uint64>(sysTimeT);
 }
 
 //float SystemTimer::FrameDelta()
